engine/sync-engine/requirements.txt
# Vector Database
qdrant-client>=1.7.0

# Embedding Model
sentence-transformers>=2.2.0
torch>=2.0.0  # Required by sentence-transformers

# Document Processing
pyyaml>=6.0
python-frontmatter>=1.0.0

# Text Chunking
langchain>=0.1.0
tiktoken>=0.5.0

# Utilities
python-dotenv>=1.0.0
colorama>=0.4.6  # For colored terminal output
